fix: eliminar todos los warnings de clippy en neuro-voice
- Convertidos loops index-based a iteradores con enumerate()
- Reemplazados arithmetic checks con saturating_sub()
- Eliminados closures redundantes en handlers
- Corregida indentación en documentación
- Agregado #[allow(clippy::too_many_arguments)] donde necesario

12 warnings eliminados: needless_range_loop, implicit_saturating_sub,
redundant_closure, unnecessary_cast, doc_lazy_continuation

+++ neuro-voice/src/audio_effects/chorus.rs
@@ -16,7 +16,7 @@ pub fn apply_chorus(audio: &[f32], sample_rate: u32, wet: f32, num_voices: usize
         let phase_offset = voice_idx as f32 * 2.0 * PI / num_voices as f32;
 
         let base_delay_samples = (sample_rate_f * base_delay_ms / 1000.0) as usize;
-        let depth_samples = (sample_rate_f * depth_ms / 1000.0) as f32;
+        let depth_samples = sample_rate_f * depth_ms / 1000.0;
 
         for i in 0..len {
             let t = i as f32 / sample_rate_f;
